requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.2
frozenlist==1.3.3
h11==0.14.0
httptools==0.5.0
idna==3.4
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
openai==0.27.7
pydantic==1.10.8
Pygments==2.15.1
python-dotenv==1.0.0
PyYAML==6.0
requests==2.31.0
rich==13.3.5
sniffio==1.3.0
sse-starlette==1.6.1
starlette==0.27.0
tqdm==4.65.0
typing_extensions==4.6.1
urllib3==2.0.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
